Deleted the themes folders and added a zip files for: • Compatible Theme Skins - Each theme folder includes a customizable theme skin. There are 34 custom themes available for osFinancials/TurboCASH. You can view screenshots of the available themes to help you make your choice.
o Installed Themes: These 7 themes are included in the osFinancials5 installations and may include fixes. They can be downloaded as a zip file and copied to the `.../bin/themes` folder of your osFinancials5 installation directory.

o  New Themes - A collection of 27 themes. These themes are available for download as a zip file. Once unzipped, you can copy the themes you need to the `.../bin/themes` folder of your osFinancials5 installation directory. These themes are not included in the osFinancials5 installs.